+++ src/equal.c
@@ -254,6 +254,7 @@ int equal_object(SEXP x, SEXP y, bool na_equal) {
   // Pointer comparison is safe for these types
   switch (type) {
   case NILSXP:
+    return x == y ? (na_equal ? true : NA_LOGICAL) : false;
   case SYMSXP:
   case SPECIALSXP:
   case BUILTINSXP:
@@ -292,7 +293,15 @@ int equal_object(SEXP x, SEXP y, bool na_equal) {
     if (eq <= 0) {
       return eq;
     }
-    eq = equal_object(CDR(x), CDR(y), na_equal);
+
+    x = CDR(x);
+    y = CDR(y);
+
+    if (x == R_NilValue && y == R_NilValue) {
+      return true;
+    }
+
+    eq = equal_object(x, y, na_equal);
     if (eq <= 0) {
       return eq;
     }